+++ src/commands/hardis/org/diagnose/audittrail.ts
@@ -31,8 +31,10 @@ Regular setup actions performed in major orgs are filtered.
 - Custom App Licenses
   - addeduserpackagelicense
   - granteduserpackagelicense
+  - revokeduserpackagelicense
 - Customer Portal
   - createdcustomersuccessuser
+  - CSPUserDisabled
 - Currency
   - updateddatedexchrate
 - Data Management
@@ -65,6 +67,8 @@ Regular setup actions performed in major orgs are filtered.
   - changedroleforuserfromnone
   - changedUserEmailVerifiedStatusUnverified
   - changedUserEmailVerifiedStatusVerified
+  - changedknowledgeuseroffon
+  - changedsupportuseroffon
   - changedUserPhoneNumber
   - changedUserPhoneVerifiedStatusUnverified
   - deactivateduser
@@ -84,6 +88,8 @@ Regular setup actions performed in major orgs are filtered.
   - PermSetLicenseUnassign
   - registeredUserPhoneNumber
   - resetpassword
+  - suNetworkAdminLogin
+  - suNetworkAdminLogout
   - suOrgAdminLogin
   - suOrgAdminLogout
   - unfrozeuser
@@ -203,8 +209,15 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
     this.allowedSectionsActions = {
       '': ['createScratchOrg', 'changedsenderemail', 'deleteScratchOrg', 'loginasgrantedtopartnerbt'],
       'Certificate and Key Management': ['insertCertificate'],
-      'Custom App Licenses': ['addeduserpackagelicense', 'granteduserpackagelicense'],
-      'Customer Portal': ['createdcustomersuccessuser'],
+      'Custom App Licenses': [
+        'addeduserpackagelicense',
+        'granteduserpackagelicense',
+        'revokeduserpackagelicense'
+      ],
+      'Customer Portal': [
+        'createdcustomersuccessuser',
+        'CSPUserDisabled'
+      ],
       Currency: ['updateddatedexchrate'],
       'Data Management': ['queueMembership'],
       'Email Administration': ['dkimRotationSuccessful', 'dkimRotationPreparationSuccessful'],
@@ -232,6 +245,8 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
         'changedpassword',
         'changedUserEmailVerifiedStatusUnverified',
         'changedUserEmailVerifiedStatusVerified',
+        'changedknowledgeuseroffon',
+        'changedsupportuseroffon',
         'changedUserPhoneNumber',
         'changedUserPhoneVerifiedStatusUnverified',
         'deactivateduser',
@@ -251,6 +266,8 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
         'PermSetLicenseUnassign',
         'registeredUserPhoneNumber',
         'resetpassword',
+        'suNetworkAdminLogin',
+        'suNetworkAdminLogout',
         'suOrgAdminLogin',
         'suOrgAdminLogout',
         'unfrozeuser',
@@ -301,6 +318,7 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
     const queryRes = await bulkQuery(auditTrailQuery, conn);
     const suspectRecords: any[] = [];
     let suspectUsers: any[] = [];
+    const suspectUsersAndActions: any = {};
     const suspectActions: any[] = [];
     const severityIconLog = getSeverityIcon('log');
     const severityIconWarning = getSeverityIcon('warning');
@@ -323,8 +341,21 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
         record.severity = 'warning';
         record.severityIcon = severityIconWarning;
         suspectRecords.push(record);
-        suspectUsers.push(record['CreatedBy.Username'] + ' - ' + record['CreatedBy.Name']);
-        suspectActions.push(`${section} - ${record.Action}`);
+        const suspectUserDisplayName = `${record['CreatedBy.Name']}`;
+        suspectUsers.push(suspectUserDisplayName);
+        const actionFullName = `${section} - ${record.Action}`;
+        suspectActions.push(actionFullName);
+        if (!suspectUsersAndActions[suspectUserDisplayName]) {
+          suspectUsersAndActions[suspectUserDisplayName] = {
+            name: record['CreatedBy.Name'],
+            actions: [],
+          };
+        }
+        const suspectUserActions = suspectUsersAndActions[suspectUserDisplayName].actions;
+        if (!suspectUserActions.includes(record.Action)) {
+          suspectUserActions.push(record.Action);
+        }
+        suspectUsersAndActions[suspectUserDisplayName].actions = suspectUserActions;
         return record;
       }
       return record;
@@ -352,7 +383,7 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
       uxLog(this, '');
       uxLog(this, c.yellow('Related users:'));
       for (const user of suspectUsers) {
-        uxLog(this, c.yellow(`- ${user}`));
+        uxLog(this, c.yellow(`- ${user}` + ' (' + suspectUsersAndActions[user].actions.join(', ') + ")"));
       }
       uxLog(this, '');
       uxLog(this, c.yellow('Related actions:'));
@@ -380,7 +411,7 @@ This command is part of [sfdx-hardis Monitoring](${CONSTANTS.DOC_URL_ROOT}/sales
       let notifDetailText = ``;
       notifDetailText += '*Related users*:\n';
       for (const user of suspectUsers) {
-        notifDetailText += `• ${user}\n`;
+        notifDetailText += `• ${user + " (" + suspectUsersAndActions[user].actions.join(', ') + ")"}\n`;
       }
       notifDetailText += '\n';
       notifDetailText += '*Related actions*:\n';